requirements.txt
Django>=4.2,<5.0
numpy>=1.24.0
scipy>=1.10.0
matplotlib>=3.6.0
pandas>=1.5.0
rdkit>=2022.9.1
requests>=2.28.0
celery>=5.2.0
redis>=4.5.0
django-filter>=23.0
djangorestframework>=3.14.0
optuna>=3.6.0
